# openapi.json
{"openapi":"3.1.0","info":{"title":"Notescape API","version":"0.1.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/classes":{"get":{"tags":["classes"],"summary":"List Classes","operationId":"list_classes_api_classes_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Response List Classes Api Classes Get"}}}}}},"post":{"tags":["classes"],"summary":"Create Class","operationId":"create_class_api_classes_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClassCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/classes/{class_id}":{"put":{"tags":["classes"],"summary":"Update Class","operationId":"update_class_api_classes__class_id__put","parameters":[{"name":"class_id","in":"path","required":true,"schema":{"type":"integer","title":"Class Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClassUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["classes"],"summary":"Delete Class","operationId":"delete_class_api_classes__class_id__delete","parameters":[{"name":"class_id","in":"path","required":true,"schema":{"type":"integer","title":"Class Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/files/{class_id}":{"get":{"tags":["files"],"summary":"List Files","operationId":"list_files_api_files__class_id__get","parameters":[{"name":"class_id","in":"path","required":true,"schema":{"type":"integer","title":"Class Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["files"],"summary":"Upload File","operationId":"upload_file_api_files__class_id__post","parameters":[{"name":"class_id","in":"path","required":true,"schema":{"type":"integer","title":"Class Id"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_file_api_files__class_id__post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/files/{file_id}":{"delete":{"tags":["files"],"summary":"Delete File","operationId":"delete_file_api_files__file_id__delete","parameters":[{"name":"file_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"File Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/contact":{"post":{"tags":["contact"],"summary":"Contact","operationId":"contact_api_contact_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContactIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/files/{file_id}/chunks":{"get":{"tags":["chunks"],"summary":"List Chunks","operationId":"list_chunks_api_files__file_id__chunks_get","parameters":[{"name":"file_id","in":"path","required":true,"schema":{"type":"string","title":"File Id"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":20,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Offset"}},{"name":"full","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Full"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","additionalProperties":true},"title":"Response List Chunks Api Files  File Id  Chunks Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/chunks":{"post":{"tags":["chunks"],"summary":"Create Chunks","description":"Creates chunks for each file id.\nUses the SAME uploads folder mounted in main.py (request.app.state.uploads_root).","operationId":"create_chunks_api_chunks_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChunkRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ChunkPreview"},"type":"array","title":"Response Create Chunks Api Chunks Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/embeddings/build":{"post":{"tags":["embeddings"],"summary":"Build Embeddings","operationId":"build_embeddings_api_embeddings_build_post","parameters":[{"name":"class_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"description":"Only build for this class","title":"Class Id"},"description":"Only build for this class"},{"name":"limit","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"default":1000,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/flashcards/ping":{"get":{"tags":["flashcards"],"summary":"Ping","operationId":"ping_api_flashcards_ping_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/flashcards/{class_id}":{"get":{"tags":["flashcards"],"summary":"List Flashcards For Class","operationId":"list_flashcards_for_class_api_flashcards__class_id__get","parameters":[{"name":"class_id","in":"path","required":true,"schema":{"type":"integer","title":"Class Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FlashcardOut"},"title":"Response List Flashcards For Class Api Flashcards  Class Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/flashcards/{card_id}":{"delete":{"tags":["flashcards"],"summary":"Delete Flashcard","operationId":"delete_flashcard_api_flashcards__card_id__delete","parameters":[{"name":"card_id","in":"path","required":true,"schema":{"type":"string","description":"flashcards.id (UUID)","title":"Card Id"},"description":"flashcards.id (UUID)"}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/flashcards/ensure-embeddings":{"post":{"tags":["flashcards"],"summary":"Ensure Embeddings","operationId":"ensure_embeddings_api_flashcards_ensure_embeddings_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EnsureEmbeddingsReq"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/flashcards/generate/":{"post":{"tags":["flashcards"],"summary":"Generate","operationId":"generate_api_flashcards_generate__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateReq"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/FlashcardOut"},"type":"array","title":"Response Generate Api Flashcards Generate  Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/flashcards/generate":{"post":{"tags":["flashcards"],"summary":"Generate","operationId":"generate_api_flashcards_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateReq"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/FlashcardOut"},"type":"array","title":"Response Generate Api Flashcards Generate Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/__routes":{"get":{"summary":"  Routes","operationId":"__routes___routes_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_upload_file_api_files__class_id__post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_file_api_files__class_id__post"},"ChunkPreview":{"properties":{"file_id":{"type":"string","title":"File Id"},"total_chunks":{"type":"integer","title":"Total Chunks"},"previews":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Previews"},"note":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Note"}},"type":"object","required":["file_id","total_chunks","previews"],"title":"ChunkPreview"},"ChunkRequest":{"properties":{"file_ids":{"items":{"type":"string"},"type":"array","title":"File Ids"},"by":{"type":"string","enum":["auto","page","chars"],"title":"By","default":"auto"},"size":{"type":"integer","title":"Size","default":2000},"overlap":{"type":"integer","title":"Overlap","default":200},"preview_limit_per_file":{"type":"integer","title":"Preview Limit Per File","default":3}},"type":"object","required":["file_ids"],"title":"ChunkRequest"},"ClassCreate":{"properties":{"name":{"type":"string","title":"Name"},"subject":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Subject"}},"type":"object","required":["name"],"title":"ClassCreate"},"ClassUpdate":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"subject":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Subject"}},"type":"object","title":"ClassUpdate"},"ContactIn":{"properties":{"name":{"type":"string","title":"Name"},"email":{"type":"string","format":"email","title":"Email"},"message":{"type":"string","title":"Message"}},"type":"object","required":["name","email","message"],"title":"ContactIn"},"EnsureEmbeddingsReq":{"properties":{"limit":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Limit","default":500}},"type":"object","title":"EnsureEmbeddingsReq"},"FlashcardOut":{"properties":{"id":{"type":"string","title":"Id"},"class_id":{"type":"integer","title":"Class Id"},"source_chunk_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Source Chunk Id"},"question":{"type":"string","title":"Question"},"answer":{"type":"string","title":"Answer"},"hint":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Hint"},"difficulty":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Difficulty","default":"medium"},"tags":{"items":{"type":"string"},"type":"array","title":"Tags","default":[]}},"type":"object","required":["id","class_id","source_chunk_id","question","answer"],"title":"FlashcardOut"},"GenerateReq":{"properties":{"class_id":{"type":"integer","title":"Class Id"},"topic":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Topic"},"top_k":{"type":"integer","maximum":100.0,"minimum":1.0,"title":"Top K","default":12},"n_cards":{"type":"integer","maximum":50.0,"minimum":1.0,"title":"N Cards","default":10}},"type":"object","required":["class_id"],"title":"GenerateReq"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}